Implemented SimExternalConnection::write
